feat(export settings): Code rearranged
