+++ packages/portal/src/components/ExportSettingsDialog/entities/export-settings-form.ts
@@ -1,59 +1,9 @@
-import { ExportedEntityKind, ExportedEntityTransformation, ExportedFileFormat } from "../api/useExport";
-import { ExportSettingsFormField, ExportSettingsFormFieldKind, ExportSettingsFormFieldOption } from "./export-settings-form-field";
+import { ExportedEntityKind } from "../api/useExport";
+import { ExportSettingsFormField, ExportSettingsFormFieldKind, ExportSettingsFormFieldOption, FIELD_LABEL_FILE_FORMAT, FIELD_LABEL_FILE_FORMAT_OAS, FIELD_LABEL_OAS_EXTENSIONS, FIELD_LABEL_SPECIFICATION_TYPE, FIELD_OPTION_LIST_FILE_FORMAT_OAS, FIELD_OPTION_LIST_OAS_EXTENSIONS, FIELD_OPTION_LIST_SPECIFICATION_TYPE } from "./export-settings-form-field";
 
 export interface ExportSettingsForm {
   fields: ExportSettingsFormField[]
 }
-
-const FIELD_LABEL_SPECIFICATION_TYPE = 'Specification type'
-const FIELD_LABEL_FILE_FORMAT_OAS = 'File format for OpenAPI specifications'
-const FIELD_LABEL_FILE_FORMAT = 'File format'
-const FIELD_LABEL_OAS_EXTENSIONS = 'OpenAPI extensions'
-
-const FIELD_OPTION_SPECIFICATION_TYPE_REDUCED: ExportSettingsFormFieldOption = {
-  label: 'Reduced source specifications',
-  value: ExportedEntityTransformation.REDUCED_SOURCE_SPECIFICATIONS
-}
-const FIELD_OPTION_SPECIFICATION_TYPE_COMBINED: ExportSettingsFormFieldOption = {
-  label: 'Combined specification',
-  value: ExportedEntityTransformation.MERGED_SPECIFICATION
-}
-const FIELD_OPTION_FILE_FORMAT_OAS_JSON: ExportSettingsFormFieldOption = {
-  label: 'JSON',
-  value: ExportedFileFormat.JSON
-}
-const FIELD_OPTION_FILE_FORMAT_OAS_YAML: ExportSettingsFormFieldOption = {
-  label: 'YAML',
-  value: ExportedFileFormat.YAML
-}
-const FIELD_OPTION_FILE_FORMAT_OAS_INTERACTIVE_HTML: ExportSettingsFormFieldOption = {
-  label: 'Interactive HTML',
-  value: ExportedFileFormat.HTML
-}
-const FIELD_OPTION_OAS_EXTENSIONS_PRESERVE: ExportSettingsFormFieldOption = {
-  label: 'Preserve all OAS extensions',
-  value: 'preserve'
-}
-const FIELD_OPTION_OAS_EXTENSIONS_REMOVE: ExportSettingsFormFieldOption = {
-  label: 'Remove OAS extensions',
-  value: 'remove'
-}
-
-const FIELD_OPTION_LIST_SPECIFICATION_TYPE: ReadonlyArray<ExportSettingsFormFieldOption> = [
-  FIELD_OPTION_SPECIFICATION_TYPE_REDUCED,
-  FIELD_OPTION_SPECIFICATION_TYPE_COMBINED,
-]
-
-const FIELD_OPTION_LIST_FILE_FORMAT_OAS: ReadonlyArray<ExportSettingsFormFieldOption> = [
-  FIELD_OPTION_FILE_FORMAT_OAS_YAML,
-  FIELD_OPTION_FILE_FORMAT_OAS_JSON,
-  FIELD_OPTION_FILE_FORMAT_OAS_INTERACTIVE_HTML,
-]
-
-const FIELD_OPTION_LIST_OAS_EXTENSIONS: ReadonlyArray<ExportSettingsFormFieldOption> = [
-  FIELD_OPTION_OAS_EXTENSIONS_PRESERVE,
-  FIELD_OPTION_OAS_EXTENSIONS_REMOVE,
-]
 
 export const EXPORT_SETTINGS_FORM_FIELDS_BY_PLACE: Record<ExportedEntityKind, ExportSettingsFormField[]> = {
   [ExportedEntityKind.VERSION]: [
